fix(linter): a crash, three false errors, and six diagnostics that never fired
A rule walking `0..setting_count` reported one diagnostic per index, and
`Linter` spread that array into `push`: a `SplineIK3D` at 130,000 settings
exceeded the argument limit and the whole file returned nothing at all, parse
errors included. `reportedIndices.ts` caps what one rule names, and `Linter`
appends one at a time so no rule can throw the file away again.

The false errors, each on a file Godot loads unaltered:

- `secret_character`/`ellipsis_char` counted UTF-16 units on a literal whose
  escapes were never decoded, so `"🔒"` and `"…"` both errored. One
  `v.singleCharacter` now counts code points off the decoded text.
- `trail_lifetime` put the hint's floor on the error tier while the setter
  refuses one epsilon below it.
- `checkResourceExists` answered false for a value that is not a reference at
  all, so `shape = "hello"` also claimed a resource named `hello` was missing.

And the silences: the bit-field combinators are INT slots and never applied the
shared `truncatedInt` (13 slots, swept by a new `truncatedInts` guard); a
`%UniqueName` inside an instanced sub-scene read as dangling; an uncataloged
ancestor read as visible in both visibility cascades; `basisColumnScales` and
`hasNonUnitScale3D` refused an `inf`/`nan` or degenerate basis Godot reads a
scale off; `visibility_range_fade_mode` was compared as raw text; `outline_size`
is `Variant::INT`, not float.

`maskedBitField` does its bit arithmetic in BigInt — `|`/`&` are int32 and the
slot is int64 — and `basisDeterminant` moves to `godot/math.ts`, where both the
linter and the renderer reach it.

+++ packages/textscene-core/src/nodes/3d/label3d/linterParser.ts
@@ -48,7 +48,9 @@ validatorRegistry.registerAll('Label3D', {
   // label_3d.cpp:155 hints "0,127,1,suffix:px" — closed at both ends, the
   // suffix being a unit and not a bound; set_outline_size:873-880 is a bare
   // assignment (only guarded against a redundant set), so both ends warn.
-  outline_size: v.float('outline_size', {
+  // `v.int`, like `font_size` below: the slot is Variant::INT, so it also owes
+  // the shared truncation warning and the unstorable-literal refusal.
+  outline_size: v.int('outline_size', {
     min: 0,
     max: 127,
     hinted: 'label_3d.cpp:155',
